feat(warmup): update warmup screen
